backend/auth/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
sqlalchemy==2.0.41
asyncpg==0.30.0
alembic==1.15.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
firebase-admin==6.8.0
prometheus-client==0.22.1
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.2.0
